+++ actionmailer/lib/action_mailer/base.rb
@@ -844,8 +844,9 @@ module ActionMailer
     def collect_responses(headers) #:nodoc:
       responses = []
 
+      template_name = headers.delete(:template_name) || action_name
       if block_given?
-        collector = ActionMailer::Collector.new(lookup_context) { render(action_name) }
+        collector = ActionMailer::Collector.new(lookup_context) { render(template_name) }
         yield(collector)
         responses = collector.responses
       elsif headers[:body]
@@ -855,9 +856,8 @@ module ActionMailer
         }
       else
         templates_path = headers.delete(:template_path) || self.class.mailer_name
-        templates_name = headers.delete(:template_name) || action_name
 
-        each_template(Array(templates_path), templates_name) do |template|
+        each_template(Array(templates_path), template_name) do |template|
           self.formats = template.formats
 
           responses << {